Consider any file with a shebang executable
Any file with a shebang (first line beginning with "#!" followed
immediately by a path to an executable) will be treated as an executable
by QuickServ, even on Windows. It will be run using the filepath as the
second or third argument to the executable specified in the shebang.

This will hopefully make it easy for people to create executables on any
OS without using shell script hacks, and without having to do any chmod
nonsense.